Roblox ResChanger
Roblox Resolution Changer:

This tool allows players who enjoy playing in "stretched" mode to automatically adjust their resolution in Roblox without having to change it manually. It automatically detects whether the program is running on a German or English system and adjusts the user interface accordingly.

Features:

Changes the resolution based on the user's preferred settings.
Automatically detects whether the program is running on a German or English system and adjusts the user interface accordingly.
Easy to use and saves time on adjusting the resolution.